// src/app/types/enum/documents-type.enum.ts
/**
 * Documents Type to be used in labelex application
 * Add other document type if necessary
 *
 * Author: Marlon Micael J. Cuevas
 * Date Created: Oct. 7, 2020
 */
export enum DocumentsType {
    NCV = 'CORRESPONDENCE_NO_COMMERCIAL_VALUE', // Correspondence/No Commercial Value
    PERSONAL = 'PERSONAL_DOCUMENT', // Personal Document
    INTEROFFICE = 'INTEROFFICE_DOCUMENT', // Interoffice Doccment
    BUSINESS = 'BUSINESS_DOCUMENT', // Business Document
    OTHER = 'OTHER', // Other
}