// src/app/types/enum/service-type.enum.ts
/**
 * Service types enum for LABELEX application.
 * Lists the available service types in LABELEX application.
 * Author: Marlon Micael J. Cuevas
 * Date Created: Nov. 24, 2020
 */
export enum ServiceType {
    INTERNATIONAL_PRIORITY = 'INTERNATIONAL_PRIORITY',
    INTERNATIONAL_ECONOMY = 'INTERNATIONAL_ECONOMY',
    INTERNATIONAL_FIRST = 'INTERNATIONAL_FIRST',
    INTERNATIONAL_PRIORITY_FREIGHT = 'INTERNATIONAL_PRIORITY_FREIGHT',
    INTERNATIONAL_ECONOMY_FREIGHT = 'INTERNATIONAL_ECONOMY_FREIGHT',
    EUROPE_FIRST_INTERNATIONAL_PRIORITY = 'EUROPE_FIRST_INTERNATIONAL_PRIORITY',
    FIRST_OVERNIGHT = 'FIRST_OVERNIGHT',
    PRIORITY_OVERNIGHT = 'PRIORITY_OVERNIGHT',
    STANDARD_OVERNIGHT = 'STANDARD_OVERNIGHT',
    FEDEX_2_DAY = 'FEDEX_2_DAY',
    FEDEX_2_DAY_AM = 'FEDEX_2_DAY_AM',
    FEDEX_EXPRESS_SAVER = 'FEDEX_EXPRESS_SAVER',
    FEDEX_GROUND = 'FEDEX_GROUND',
    INTERNATIONAL_GROUND = 'INTERNATIONAL_GROUND',
    SAME_DAY = 'SAME_DAY',
    SAME_DAY_CITY = 'SAME_DAY_CITY'
}